Added Swagger and updated Readme file
