+++ README.md
@@ -1,2 +1,15 @@
 # my-shop-backend
-Serverless backend to serve FE repo shop-react-redux-cloudfront
+**Serverless backend to serve FE repo shop-react-redux-cloudfront**
+
+
+[![AWS Lambda](https://img.shields.io/badge/AWS-Lambda-orange)](https://docs.aws.amazon.com/lambda/?icmpid=docs_homepage_featuredsvcs) [![AWS API Gateway](https://img.shields.io/badge/AWS-API_Gateway-orange)](https://docs.aws.amazon.com/apigateway/?icmpid=docs_homepage_serverless) [![Serverless framework](https://img.shields.io/badge/Serverless_framework-red)](https://www.serverless.com/) [![Node.js](https://img.shields.io/badge/Node.js-green)](https://nodejs.org) [![Typescript](https://img.shields.io/badge/Typescript-blue)](https://www.typescriptlang.org/)
+
+
+## Endpoints
+- [Swagger Documentation](https://rdw48vy1ca.execute-api.eu-west-1.amazonaws.com/swagger)
+- [Products](https://xxpvdpuwni.execute-api.eu-west-1.amazonaws.com/dev/products)
+- [Products/{id}](https://xxpvdpuwni.execute-api.eu-west-1.amazonaws.com/dev/products/7567ec4b-b10c-48c5-9345-fc73c48a80a1)[^1]
+
+
+
+[^1]: The id '7567ec4b-b10c-48c5-9345-fc73c48a80a1' is used as demo in the url.
