+++ api/types/index.ts
@@ -1,6 +1,1 @@
-export type Product = {
-    id: string | undefined;
-    title: string;
-    description: string;
-    price: number;
-}
+export type { Product, Products } from './products';
