+++ api/types/products.ts
@@ -1,0 +1,8 @@
+export type Product = {
+    id: string | undefined;
+    title: string;
+    description: string;
+    price: number;
+}
+
+export type Products = Product[];
